Move failover code from module_utils/robot to module_utils/failover.
