+++ plugins/modules/failover_ip_info.py
@@ -88,6 +88,8 @@ server_number:
 from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.community.hrobot.plugins.module_utils.robot import (
     ROBOT_DEFAULT_ARGUMENT_SPEC,
+)
+from ansible_collections.community.hrobot.plugins.module_utils.failover import (
     get_failover_record,
     get_failover_state,
 )
